Finalized Cuboid, still need someone that can fix the algorithms for some types. Fixed Sinjai derp

+++ Commands/CmdTest.cs
@@ -20,8 +20,8 @@ namespace CommandDll
 		string _Author = "Merlin33069";
 		public string Author { get { return _Author; } }
 
-		int _Version = 1;
-		public int Version { get { return _Version; } }
+		decimal _Version = 1.00m;
+		public decimal Version { get { return _Version; } }
 
 		string _CUD = "";
 		public string CUD { get { return _CUD; } }
