+++ Commands/Information/CmdHelp.cs
@@ -34,8 +34,8 @@ namespace CommandDll.Information {
 		string _Author = "Nerketur";
 		public string Author { get { return _Author; } }
 
-		int _Version = 1;
-		public int Version { get { return _Version; } }
+		decimal _Version = 1.00m;
+		public decimal Version { get { return _Version; } }
 
 		string _CUD = "";
 		public string CUD { get { return _CUD; } }
